% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externallinks.R
\name{add.multimir.links}
\alias{add.multimir.links}
\title{Add External Database Link for Each of the multiMiR Result Entry}
\usage{
add.multimir.links(x, org)
}
\arguments{
\item{x}{PLACEHOLDER}

\item{org}{PLACEHOLDER}
}
\description{
This is an internal multiMiR function that is not intended to be used
directly.  Please use \code{get.multimir}.
}
