% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabletypes.R
\name{all_tables}
\alias{all_tables}
\alias{all_tables,}
\alias{validated_tables,}
\alias{predicted_tables,}
\alias{associations_tables,}
\alias{tables_wo_target}
\alias{validated_tables}
\alias{predicted_tables}
\alias{associations_tables}
\alias{tables_wo_target}
\alias{conserved_tables}
\alias{get_score_var}
\title{Functions defining the category each table belongs to.}
\usage{
all_tables()

validated_tables()

predicted_tables()

associations_tables()

tables_wo_target()

conserved_tables()

get_score_var(.table)
}
\description{
One of three types: predicted, validated, or associations (disease/drug).
Additionally two functions define characteristics of tables: those without a
target column \code{tables_wo_target} and those with conserved target sites
\code{conserved_tables}.
}
\keyword{tables}
