% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmquery.R
\name{as.mmquery}
\alias{as.mmquery}
\alias{print.mmquery}
\title{S3 constructor and methods for object returned by \code{get_multimir()}.}
\usage{
as.mmquery(a_list)

\method{print}{mmquery}(x)
}
\value{
An \code{mmquery} object.
}
\description{
This package's primary user-facing object. Contains the SQL statement and the
returned data query, as well as a summary table depending on
specified option.
}
\keyword{internal}
