% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmquery.R
\name{as.mmquery_bioc}
\alias{as.mmquery_bioc}
\title{S4 constructor and methods for object returned by \code{get_multimir()}.}
\usage{
as.mmquery_bioc(a_list)
}
\value{
An \code{mmquery_bioc} object.
}
\description{
This package's primary user-facing object. Contains the SQL statement and the
returned data query, as well as a summary table depending on
specified option.
}
\keyword{internal}
