% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlclasses.R
\name{as_mmsql}
\alias{as_mmsql}
\alias{as_mmsql,}
\alias{as_where_list,}
\alias{as_where,}
\alias{as_orderby,}
\alias{is_where_list}
\title{S3 Class constructors for objects defining SQL query parts
and a collection of these parts (\code{mmsql}).}
\usage{
as_mmsql(.select = NULL, .from = NULL, .on = NULL, .where_list = NULL,
  .orderby = NULL)
}
\description{
The collection object has a defined set of components that match the multiMiR
database and defined options in \code{get.multimir()}. Conceptually this is
split into two parts, the relatively straightforward SELECT, FROM, and ON
portion of the query and the more complex filtering and sorting operations:
WHERE and ORDER BY.  The latter have their own classes, the former are
resolved as strings (or character vectors) in the functions defining handling
of each sql table (\code{sql_} prefix).
}
\keyword{internal}
