% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlclasses.R
\name{as_mmsql_components}
\alias{as_mmsql_components}
\alias{as_mmsql_components,}
\alias{as_where_list,}
\alias{as_where,}
\alias{as_orderby,}
\alias{is_where_list}
\alias{as_where_list}
\alias{as_where}
\alias{is_where}
\alias{as_orderby}
\title{S3 Class constructors for objects defining SQL query components
and a collection of these parts (\code{mmsql_components}).}
\usage{
as_mmsql_components(.select = NULL, .from = NULL, .on = NULL,
  .where_list = NULL, .orderby = NULL, typeattr = NULL)

as_where_list(...)

as_where(.vars, .connect = NULL, .operator, .value = "\%s")

is_where(x)

as_orderby(.vars, .order)
}
\value{
\code{as_mmsql_components}: A collection of components that make up a
SQL query. 
\code{as_where_list}, \code{as_where}, \code{as_orderby}: Individual
components of a SQL query.
}
\description{
The collection object has a defined set of components that match the multiMiR
database and defined options in \code{get_multimir()}. Conceptually this is
split into two parts, the relatively straightforward SELECT, FROM, and ON
portion of the query and the more complex filtering and sorting operations:
WHERE and ORDER BY.  The latter have their own classes, the former are
resolved as strings (or character vectors) in the functions defining handling
of each sql table (\code{sql_} prefix).
}
\keyword{internal}
