% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildquery.R
\name{build_mmsql}
\alias{build_mmsql}
\alias{expand_query}
\alias{expand_select}
\alias{expand_from}
\alias{expand_on}
\alias{expand_where_list}
\alias{expand_where}
\alias{expand_orderby}
\alias{expand_limit}
\alias{merge_order}
\title{Constructors for parts of SQL queries
Expand_query converts a \code{mmyquery} object to a SQL query string}
\usage{
build_mmsql(.table, org, mirna = NULL, target = NULL, disease.drug = NULL,
  predicted.site = NULL, predicted.cutoff.type = NULL,
  predicted.cutoff = NULL, limit = NULL)

expand_query(x)

expand_select(x)

expand_from(x)

expand_on(x)

expand_where_list(x)

expand_where(x)

expand_orderby(x)

expand_limit(x)

merge_order(.list)
}
\value{
A complete SQL statement and related information.
}
\description{
Constructors for parts of SQL queries
Expand_query converts a \code{mmyquery} object to a SQL query string
}
\keyword{internal}
