% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{check_tables}
\alias{check_tables}
\title{Check table argument matches valid table name}
\usage{
check_tables(table, force_update = FALSE)
}
\description{
Check table argument matches valid table name
}
\keyword{internal}
