% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmquery.R
\docType{methods}
\name{columns,mmquery_bioc-method}
\alias{columns,mmquery_bioc-method}
\title{S4 methods for mmquery_bioc class - based on AnnotationDbi accessor methods}
\usage{
\S4method{columns}{mmquery_bioc}(x)
}
\arguments{
\item{x}{An mmquery_bioc object.}

\item{keys}{A result of the keys() function. For the mmquery_bioc class this
is a character vector of microRNA's in the returned mmquery_bioc object.}

\item{keytype}{allows the user to discover which keytypes can be passes in to
select or keys  and the keytype argument}

\item{columns}{lists the columns that can be returned for the
\code{mmquery_bioc} object.}

\item{...}{additional arguments}
}
\description{
S4 methods for mmquery_bioc class - based on AnnotationDbi accessor methods
}
