% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{default_cutoff}
\alias{default_cutoff}
\title{If null, set default predicted.cutoff}
\usage{
default_cutoff(predicted.cutoff.type, predicted.cutoff)
}
\value{
The default cutoff value.
}
\description{
If null, set default predicted.cutoff
}
\keyword{internal}
