% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deprecate_arg}
\alias{deprecate_arg}
\title{Internal function for sending deprecation messages}
\usage{
deprecate_arg(name = c("url", "schema.file", "db.tables", "cutoff.file"))
}
\arguments{
\item{name}{One of several predefined arguments that are being deprecated.
All are URLs or URL paths now set by package/global options.}
}
\description{
Internal function for sending deprecation messages
}
\keyword{internal}
