% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmquery.R
\name{extract_mmquery}
\alias{extract_mmquery}
\title{Creates all objects needed for the legacy S3 return object and the new S4
object.}
\usage{
extract_mmquery(outlist, org, .args, summary = FALSE, use.tibble = FALSE)
}
\value{
A list of data queried, summary of results, and related input
  parameters.

a list for packaging by \code{as.mmquery} and \code{as.mmquery_bioc}
}
\description{
Creates all objects needed for the legacy S3 return object and the new S4
object.
}
\keyword{internal}
