% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{get.multimir.by.table}
\alias{get.multimir.by.table}
\title{Get microRNA/target Information from a Given Table in the multiMiR Package}
\usage{
get.multimir.by.table(table = NULL, url = NULL, org = c("hsa", "mmu",
  "rno"), mirna = NULL, target = NULL, disease.drug = NULL,
  predicted.cutoff = NULL, predicted.cutoff.type = "p",
  predicted.site = "conserved", mirna.table = "mirna",
  target.table = "target")
}
\description{
This is an internal multiMiR function that is not intended to be used
directly.  Please use \code{get.multimir}.
}
