\name{get.multimir.cutoffs}

\alias{get.multimir.cutoffs}

\title{
  Load Pre-calculated Prediction Score Cutoffs in the multiMiR Package
}

\description{
  This is an internal multiMiR function that is not intended to be
  used directly.
  Please set prediction score cutoff in \code{get.multimir}.
}

