% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{get.multimir.disease}
\alias{get.multimir.disease}
\title{Get microRNA-disease/drug Relationships from the multiMiR Package}
\usage{
get.multimir.disease(url = NULL, org = "hsa", mirna = NULL,
  target = NULL, disease.drug = NULL)
}
\description{
This is an internal multiMiR function that is not intended to be used
directly.  Please use \code{get.multimir}.
}
