% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{get.multimir.predicted}
\alias{get.multimir.predicted}
\title{Get Predicted microRNA-target Interactions from the multiMiR Package}
\usage{
get.multimir.predicted(url = NULL, org = "hsa", mirna = NULL,
  target = NULL, cutoff = NULL, cutoff.type = "p", site = "conserved")
}
\description{
This is an internal multiMiR function that is not intended to be used
directly.  Please use \code{get.multimir}.
}
