% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{get.multimir.validated}
\alias{get.multimir.validated}
\title{Get Validated microRNA-target Interactions from the multiMiR Package}
\usage{
get.multimir.validated(url = NULL, org = "hsa", mirna = NULL,
  target = NULL)
}
\description{
This is an internal multiMiR function that is not intended to be used
directly.  Please use \code{get.multimir}.
}
