% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiMiR_package.R
\docType{package}
\name{multiMiR}
\alias{multiMiR}
\alias{multimir}
\alias{multiMiR-package}
\title{MultiMiR: R package for accessing the multiMiR database}
\description{
This package provides an interface to the multiMiR database of
microRNA-target interactions, and disease and drug associations. See
\url{http://multimir.org} and the vignette ('multiMiR') for more
details.
}
\references{
[Add reference here]
}
