% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{multimir.summary}
\alias{multimir.summary}
\title{Summarize microRNA/target Information from the multiMiR Package}
\usage{
multimir.summary(result, pair.index = 2:6, order.by = "all.sum")
}
\arguments{
\item{result}{PLACEHOLDER}

\item{pair.index}{PLACEHOLDER}

\item{order.by}{PLACEHOLDER}
}
\description{
This is an internal multiMiR function that is not intended to be used
directly.  Please use \code{get.multimir}.
}
\keyword{internal}
