% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{myurlencode}
\alias{myurlencode}
\alias{myurlencode}
\title{Encode a URL Before Submitting It to the multiMiR Web Server}
\usage{
myurlencode(url)

myurlencode(url)
}
\description{
This is an internal multiMiR function that is not intended to be used
directly.  Please use \code{get.multimir}.

This is an internal multiMiR function that is not intended to be used
directly.  Please use \code{get.multimir}.
}
