% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad}
\alias{pad}
\title{Pad single space on each side of an input}
\usage{
pad(x)
}
\value{
Input value wrapped in single spaces.
}
\description{
Pad single space on each side of an input
}
\keyword{internal}
