% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parens_quote}
\alias{parens_quote}
\title{Prep certain names for use in SQL query by adding parens}
\usage{
parens_quote(x)
}
\value{
The input value wrapped in quotes and then parentheses.
}
\description{
Prep certain names for use in SQL query by adding parens
}
\keyword{internal}
