% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parens_wrap}
\alias{parens_wrap}
\title{Collapse a vector to a single comma-separated string and wrap in parentheses}
\usage{
parens_wrap(x)
}
\description{
Collapse a vector to a single comma-separated string and wrap in parentheses
}
\keyword{internal}
