% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{parse.multimir}
\alias{parse.multimir}
\title{Parse the Result Returned by the multiMiR Web Server}
\usage{
parse.multimir(HTML.result)
}
\description{
This is an internal multiMiR function that is not intended to be used
directly.  Please use \code{get.multimir}.
}
