% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{parse_orgs}
\alias{parse_orgs}
\title{Each org can be specified in one of 3 ways -- this standardizes the argument
into the 3 char abbreviation.}
\usage{
parse_orgs(org)
}
\value{
A standardized, abbreviated form of the input org.
}
\description{
Each org can be specified in one of 3 ways -- this standardizes the argument
into the 3 char abbreviation.
}
\keyword{internal}
