% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{parse_response}
\alias{parse_response}
\title{Parse the Result Returned by the multiMiR Web Server}
\usage{
parse_response(HTML.response)
}
\value{
The queried table portion of the HTML response.
}
\description{
This is an internal multiMiR function that is not intended to be used
directly.  Please use \code{get_multimir}.
}
\keyword{internal}
