% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{query.multimir}
\alias{query.multimir}
\title{Wrapper for search.multimir for adding feature (printing notification to
console)}
\usage{
\method{query}{multimir}(x)
}
\description{
Wrapper for search.multimir for adding feature (printing notification to
console)
}
\keyword{internal}
