% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querybuilders.R
\name{query_disease}
\alias{query_disease}
\title{Prepare query for miRNA-disease/drug table}
\usage{
query_disease(table = c("mir2disease", "pharmaco_mir", "phenomir"),
  mirna = NULL, target = NULL, org, mirna.table, target.table, disease.drug,
  ...)
}
\description{
This is an internal multiMiR function that is not intended to be used
directly.  Please use \code{get.multimir}.
}
