% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{query_multimir}
\alias{query_multimir}
\title{Wrapper for search_multimir for adding feature (printing notification to
console)}
\usage{
query_multimir(x, org, add.link, use.tibble)
}
\value{
The queried multimir data with the addition of a requested feature.
}
\description{
Wrapper for search_multimir for adding feature (printing notification to
console)
}
\keyword{internal}
