% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querybuilders.R
\name{query_predicted}
\alias{query_predicted}
\title{Prepare query for predicted target table}
\usage{
query_predicted(table = c("diana_microt", "elmmo", "microcosm", "miranda",
  "mirdb", "pictar", "pita", "targetscan"), mirna = NULL, target = NULL,
  org, mirna.table, target.table, predicted.cutoff, predicted.cutoff.type,
  predicted.site = c("conserved", "nonconserved", "all"), ...)
}
\description{
This is an internal multiMiR function that is not intended to be used
directly.  Please use \code{get.multimir}.
}
