% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{query_validated}
\alias{query_validated}
\title{Prepare query for validated target table}
\usage{
query_validated(table = c("mirecords", "mirtarbase", "tarbase"), mirna,
  target, org, mirna.table, target.table)
}
\arguments{
\item{table}{One of the three validated tables}
}
\description{
This is an internal multiMiR function that is not intended to be used
directly.  Please use \code{get.multimir}.
}
\keyword{internal}
