% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quote_wrap}
\alias{quote_wrap}
\title{Internal function for adding single quotes around a string}
\usage{
quote_wrap(x)
}
\arguments{
\item{x}{a string to be wrapped in single quotes.}
}
\value{
The input wrapped in single quotes.
}
\description{
Internal function for adding single quotes around a string
}
\keyword{internal}
