% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{remove_empty_strings}
\alias{remove_empty_strings}
\title{Remove empty strings from character vector.}
\usage{
remove_empty_strings(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A character vector with empty strings removed
}
\description{
The WHERE clauses for target and mirna use allow for multiple arguments
always separated by 'OR' and several columns are checked for each value 
(mirna id, acc; target symbol, entrez, ensemble).  If empty strings "" are
present in the get_multimir arguments, Targets and miRNA with empty values in
one of these columns will be incorrectly returned.  -- thus purge empty
strings first.
}
\keyword{internal}
