% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_by}
\alias{split_by}
\title{Split, order and sort lists by their components.}
\usage{
split_by(.x, .f, ...)
}
\arguments{
\item{.x}{A list or atomic vector.}

\item{.f}{A function, formula, or atomic vector.}

\item{...}{Additional arguments passed on to \code{.f}.}
}
\value{
A list split by \code{.f}
}
\description{
Copied from purrr:v0.2.2
}
\keyword{internal}
