% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlclauses.R
\name{sql_org}
\alias{sql_org}
\alias{sql_org,}
\alias{where_org,}
\alias{where_associations,}
\alias{where_conserved,}
\alias{where_cutoff}
\alias{create_cutoff_name}
\alias{cutoff_to_score}
\alias{where_org}
\alias{where_associations}
\alias{where_conserved}
\alias{where_cutoff}
\alias{create_cutoff_name}
\alias{cutoff_to_score}
\title{Generate mmsql objects for each of the three types of tables, as well as the
mirna and target tables.}
\usage{
sql_org(.table, org)

where_org(.table, org)

where_associations(.table, disease.drug)

where_conserved(.table, org, predicted.site)

where_cutoff(.table, score_var, score_cutoff)

create_cutoff_name(.table, org, predicted.site)

cutoff_to_score(.table, cutoff_name, predicted.cutoff.type, predicted.cutoff)
}
\description{
The three types of tables are predicted, validated, and associations
(disease/drug).
}
\keyword{associations}
\keyword{disease}
\keyword{drug}
\keyword{internal}
\keyword{predicted}
\keyword{tables}
\keyword{types}
\keyword{validated}
