% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{submit_request}
\alias{submit_request}
\title{General workhorse function for submitting and returning queries}
\usage{
submit_request(url = full_url("multimir.queries"), query, ...)
}
\value{
Table requested in \code{query}.
}
\description{
This is an internal multiMiR function that is not intended to be used
directly.  Please use \code{get_multimir}.
}
\keyword{internal}
