% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querybuilders.R
\name{subquery_conserved}
\alias{subquery_conserved}
\title{WHERE: conserved (predicted.site) portion of query}
\usage{
subquery_conserved(predicted.site, table, org)
}
\arguments{
\item{predicted.site}{see ?get.multimir}

\item{table}{see ?get.multimir}

\item{org}{see ?get.multimir}
}
\description{
Process conserved/nonconserved sites in miranda, pita & targetscan
(other tables don't have conserved/nonconserved annotation)
  - 2 objects created: 'name' and 'query'
  - subquery_conserved structure: conserved_var + operator + cut_value
}
\keyword{internal}
