% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querybuilders.R
\name{subquery_org}
\alias{subquery_org}
\title{Internal function for building org portion of query}
\usage{
subquery_org(table, org)
}
\arguments{
\item{table}{One of the validate, predict, or disease tables}

\item{org}{One of the 3 organisms (hsa, mmu, rno)}
}
\description{
Internal function for building org portion of query
}
\keyword{internal}
