% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{table_query_lookup}
\alias{table_query_lookup}
\title{Create/filter lookup table for all SQL tables to query}
\usage{
table_query_lookup(tbl_arg, mirna, target)
}
\description{
This is an internal multiMiR function that is not intended to be used
directly.  Please use \code{get.multimir}.
}
\keyword{internal}
