% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmquery.R
\name{as_mmquery}
\alias{as_mmquery}
\title{Constructor and methods for object returned by \code{get.multimir()}.}
\usage{
as_mmquery()
}
\description{
This package's primary user-facing object. Contains the SQL statement and the
returned data query, as well as summary and link objects depending on
specified options
}
\keyword{internal}
