% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deprecate_arg}
\alias{deprecate_arg}
\title{Internal function for sending deprecation messages}
\usage{
deprecate_arg(name = c("url", "schema.file", "db.tables", "cutoff.file"))
}
\arguments{
\item{name}{Name of a deprecated function argument.}
}
\value{
A message indicating deprecated arg and new version.
}
\description{
Internal function for sending deprecation messages
}
\keyword{internal}
