% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.multimir.cutoffs}
\alias{get.multimir.cutoffs}
\title{Load Pre-calculated Prediction Score Cutoffs in the multiMiR Package}
\usage{
get.multimir.cutoffs(cutoff.file = getOption("multimir.cutoffs.url"))
}
\description{
This is an internal multiMiR function that is not intended to be used
directly.  Please set prediction score cutoff in \code{get.multimir}.
}
