% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.multimir.cutoffs}
\alias{get.multimir.cutoffs}
\title{Load Pre-calculated Prediction Score Cutoffs in the multiMiR Package}
\usage{
get.multimir.cutoffs(name = NULL, cutoff.file = NULL)
}
\arguments{
\item{cutoff.file}{Deprecated. Set path to cutoffs file with the global
option \code{multimir.cutoffs}.}
}
\value{
Cutoff values object from remote database.
}
\description{
This is an internal multiMiR function that is not intended to be used
directly.  Please set prediction score cutoff in \code{get_multimir}.
}
\keyword{internal}
