% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parens_wrap}
\alias{parens_wrap}
\title{Prep certain names for use in SQL query by adding parens}
\usage{
parens_wrap(x)
}
\description{
Prep certain names for use in SQL query by adding parens
}
\keyword{internal}
