% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{subquery_mirnatarget}
\alias{subquery_mirnatarget}
\title{Internal function for building mirna/target portion of query}
\usage{
subquery_mirnatarget(mirna = NULL, target = NULL)
}
\arguments{
\item{mirna}{see ?get.multimir}

\item{target}{see ?get.multimir}
}
\description{
Internal function for building mirna/target portion of query
}
\keyword{internal}
